requirements/development.txt
-r base.txt

# Development tools
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Debugging
flask-debugtoolbar==0.13.1